# thoughty-server/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              },
              "example": {
                "email": "user@example.com",
                "password": "Password123!",
                "username": "johndoe",
                "website": "example"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                },
                "example": {
                  "user": {
                    "example": "example"
                  },
                  "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9",
                  "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "409": {
            "description": "Email already registered"
          }
        },
        "summary": "Register a new user",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              },
              "example": {
                "identifier": "user@example.com",
                "password": "Password123!",
                "website": "example"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login tokens, or an email-code challenge when two-factor authentication is enabled",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/AuthResponseDto"
                    },
                    {
                      "$ref": "#/components/schemas/TwoFactorChallengeResponseDto"
                    }
                  ]
                },
                "example": {
                  "user": {
                    "example": "example"
                  },
                  "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9",
                  "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9",
                  "twoFactorRequired": true,
                  "challengeToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9",
                  "expiresInSeconds": 600
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "summary": "Login with email/username and password",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/two-factor/verify": {
      "post": {
        "operationId": "AuthController_verifyTwoFactor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TwoFactorCodeDto"
              },
              "example": {
                "challengeToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9",
                "code": "123456"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                },
                "example": {
                  "user": {
                    "example": "example"
                  },
                  "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9",
                  "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
                }
              }
            }
          },
          "401": {
            "description": "Invalid, expired, or already-used challenge"
          }
        },
        "summary": "Verify an email code and finish password login",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/two-factor/resend": {
      "post": {
        "operationId": "AuthController_resendTwoFactor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TwoFactorResendDto"
              },
              "example": {
                "challengeToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Replacement single-use challenge",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TwoFactorChallengeResponseDto"
                },
                "example": {
                  "twoFactorRequired": true,
                  "challengeToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9",
                  "expiresInSeconds": 600
                }
              }
            }
          }
        },
        "summary": "Replace an active challenge and send a new email code",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/two-factor/status": {
      "get": {
        "operationId": "AuthController_twoFactorStatus",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current two-factor status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TwoFactorStatusResponseDto"
                },
                "example": {
                  "enabled": true,
                  "available": true,
                  "emailVerified": true
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get two-factor availability and current status",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/two-factor/setup": {
      "post": {
        "operationId": "AuthController_setupTwoFactor",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Single-use setup challenge",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TwoFactorChallengeResponseDto"
                },
                "example": {
                  "twoFactorRequired": true,
                  "challengeToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9",
                  "expiresInSeconds": 600
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Email a code to start enabling two-factor authentication",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/two-factor/enable": {
      "post": {
        "operationId": "AuthController_enableTwoFactor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TwoFactorCodeDto"
              },
              "example": {
                "challengeToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9",
                "code": "123456"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Two-factor authentication enabled"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Confirm the setup code and enable two-factor authentication",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/two-factor/disable": {
      "post": {
        "operationId": "AuthController_disableTwoFactor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisableTwoFactorDto"
              },
              "example": {
                "password": "Password123!"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Two-factor authentication disabled"
          },
          "401": {
            "description": "Invalid password"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Disable two-factor authentication after password confirmation",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/oauth": {
      "post": {
        "operationId": "AuthController_oauth",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthDto"
              },
              "example": {
                "provider": "google",
                "providerId": "123456789",
                "email": "user@example.com",
                "name": "John Doe",
                "avatarUrl": "https://example.com/avatar.jpg"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OAuth authentication successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                },
                "example": {
                  "user": {
                    "example": "example"
                  },
                  "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9",
                  "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields"
          }
        },
        "summary": "Authenticate with OAuth provider",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              },
              "example": {
                "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New access token"
          },
          "401": {
            "description": "Invalid or expired refresh token"
          }
        },
        "summary": "Refresh access token",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              },
              "example": {
                "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged out successfully"
          }
        },
        "summary": "Logout and invalidate refresh token",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/sessions": {
      "get": {
        "operationId": "AuthController_listSessions",
        "parameters": [
          {
            "name": "x-refresh-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Refresh-Token",
            "in": "header",
            "description": "Current refresh token for marking this session",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Active sessions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionResponseDto"
                  }
                },
                "example": [
                  {
                    "id": 42,
                    "current": true,
                    "createdAt": "2026-06-21T10:00:00.000Z",
                    "expiresAt": "2026-06-28T10:00:00.000Z"
                  }
                ]
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List active sessions for the current user",
        "tags": [
          "Authentication"
        ]
      },
      "delete": {
        "operationId": "AuthController_revokeOtherSessions",
        "parameters": [
          {
            "name": "x-refresh-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Refresh-Token",
            "in": "header",
            "description": "Current refresh token to keep active",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Other sessions revoked"
          },
          "400": {
            "description": "Current refresh token missing"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke all other active sessions",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/sessions/{id}": {
      "delete": {
        "operationId": "AuthController_revokeSession",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "x-refresh-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Refresh-Token",
            "in": "header",
            "description": "Current refresh token to protect it from revocation",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session revoked"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke one active session",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "AuthController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                },
                "example": {
                  "id": 1,
                  "username": "Daily Journal",
                  "email": "user@example.com",
                  "avatarUrl": "https://thoughty.example.com/callback",
                  "authProvider": "google_drive",
                  "emailVerified": true,
                  "twoFactorEnabled": true,
                  "createdAt": "2026-06-24T10:00:00.000Z"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user info",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/change-password": {
      "post": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              },
              "example": {
                "currentPassword": "Password123!",
                "newPassword": "Password123!"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully"
          },
          "401": {
            "description": "Unauthorized or incorrect current password"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change user password",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              },
              "example": {
                "email": "user@example.com"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reset email sent if account exists"
          }
        },
        "summary": "Request password reset email",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              },
              "example": {
                "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9",
                "newPassword": "Password123!"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successfully"
          },
          "400": {
            "description": "Invalid or expired token"
          }
        },
        "summary": "Reset password with token",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/verify-email": {
      "post": {
        "operationId": "AuthController_verifyEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailDto"
              },
              "example": {
                "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email verified successfully"
          },
          "400": {
            "description": "Invalid or expired token"
          }
        },
        "summary": "Verify email with token",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/resend-verification-email": {
      "post": {
        "operationId": "AuthController_resendVerificationEmail",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Verification email sent if needed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Resend the current user email verification message",
        "tags": [
          "Authentication"
        ]
      }
    },
...
  },
  "info": {
    "title": "Thoughty Journal API",
    "description": "API for Managing Journal Entries",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT access token"
      }
    },
    "schemas": {
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "type": "object"
          },
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "user",
          "accessToken",
          "refreshToken"
        ],
        "example": {
          "user": {
            "example": "example"
          },
          "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9",
          "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
        }
      },
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "Password123!"
          },
          "username": {
            "type": "string",
            "example": "johndoe"
          },
          "website": {
            "type": "string",
            "description": "Hidden bot-trap field; must be left empty by real users"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "example": {
          "email": "user@example.com",
          "password": "Password123!",
          "username": "johndoe",
          "website": "example"
        }
      },
...
    }
  }
}